Updated code for refactoring ES
