+++ scripts/search/mapping.py
@@ -523,10 +523,49 @@ mapping = {
                             }
 
                         }
+                   },
+                "title": {
+                        "type": "string",
+                        "fields": {
+                            "raw": {
+                                "type": "string",
+                                "index": "not_analyzed"
+                            },
+                            "symbol": {
+                                "type": "string",
+                                "analyzer": "symbols"
+                            }
+
+                        }
+                   },
+                "categories": {
+                        "type": "string",
+                        "fields": {
+                            "raw": {
+                                "type": "string",
+                                "index": "not_analyzed"
+                            },
+                            "symbol": {
+                                "type": "string",
+                                "analyzer": "symbols"
+                            }
+
+                        }
+                   },
+                "tags": {
+                        "type": "string",
+                        "fields": {
+                            "raw": {
+                                "type": "string",
+                                "index": "not_analyzed"
+                            },
+                            "symbol": {
+                                "type": "string",
+                                "analyzer": "symbols"
+                            }
+
+                        }
                    }
-
-
-
                 }
             }
         }
